a few renaming and moving
